backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
sqlmodel==0.0.8
python-multipart==0.0.5
pydantic==1.10.12
python-dotenv==1.0.0
Pillow==10.1.0
